Market Orders now work
- Region Spinner now works
- Market buy and sell orders now work
- Buy orders are ordered in asending order based on price
- Sell orders are ordered in descending order based on price
- Info form is now better formatted
- Info form shows icon
- Info form shows volume and portion size

+++ eveapi/src/main/java/com/w9jds/eveapi/Models/MarketItemBase.java
@@ -2,18 +2,20 @@ package com.w9jds.eveapi.Models;
 
 import com.google.gson.annotations.SerializedName;
 
+import java.math.BigInteger;
+
 /**
  * Created by Jeremy Shore on 2/21/16.
  */
 public class MarketItemBase {
 
     @SerializedName("id")
-    private int id;
+    private long id;
 
     @SerializedName("name")
     private String name;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
@@ -21,7 +23,7 @@ public class MarketItemBase {
         return name;
     }
 
-    void setId(int id) {
+    void setId(long id) {
         this.id = id;
     }
